Move cross compiling related code to `cross_compile` module
